+++ CHANGELOG.md
@@ -12,8 +12,9 @@ CHANGELOG for FlatCAM Evo beta
 - in Excellon Editor, Slot Plugin, made sure that once the Editor is active any change for this plugin parameters will be kept until editor is exit or that parameter is changed again
 - in Excellon Editor, Drill Array Plugin, upgraded the UI and work in progress for the features
 - prettified some GUI elements (some of those that are used in the popup messages)
-- in Excellon Editor, finished th upgrade work on the Drill Array Plugin
-- in Excellon Editor, finished th upgrade work on the Slot Array Plugin
+- in Excellon Editor, finished the upgrade work on the Drill Array Plugin
+- in Excellon Editor, finished the upgrade work on the Slot Array Plugin
+- in Excellon Editor, finished the upgrade work on the Resize Plugin
 
 29.04.2022